state history plugin: cleanup
